add entity route to entity interfaces
