--- app/src/app/shared/models/entityicon.enum.ts
@@ -1,10 +1,0 @@
-
-export enum EntityIcon {
-    ARTIST = 'palette',
-    ARTWORK = 'user',
-    GENRE = 'tags',
-    LOCATION = 'archway',
-    MATERIAL = 'scroll',
-    MOVEMENT = 'wind',
-    MOTIF = 'image'
-}
