--- app/src/app/shared/models/entitytype.enum.ts
@@ -1,10 +1,0 @@
-
-export enum EntityType {
-    ARTIST = 'artist',
-    ARTWORK = 'artwork',
-    GENRE = 'genre',
-    LOCATION = 'location',
-    MATERIAL = 'material',
-    MOVEMENT = 'movement',
-    MOTIF = 'motif'
-}
