+++ app/src/app/shared/models/genre.interface.ts
@@ -1,8 +1,7 @@
-import { Entity } from './entity.interface';
-import { EntityIcon } from './entityicon.enum';
-import { EntityType } from './entitytype.enum';
+import { Entity, EntityType, EntityIcon, EntityRoute } from './entity.interface';
 
 export interface Genre extends Entity {
     type: EntityType.GENRE;
     icon: EntityIcon.GENRE;
+    route: EntityRoute.GENRE;
 }
