+++ app/src/app/shared/models/location.interface.ts
@@ -1,6 +1,4 @@
-import { Entity } from './entity.interface';
-import { EntityType } from './entitytype.enum';
-import { EntityIcon } from './entityicon.enum';
+import { Entity, EntityType, EntityIcon, EntityRoute } from './entity.interface';
 
 export interface Location extends Entity {
     country?: string;
@@ -10,4 +8,5 @@ export interface Location extends Entity {
     lon?: string;
     type: EntityType.LOCATION;
     icon: EntityIcon.LOCATION;
+    route: EntityRoute.LOCATION;
 }
