+++ app/src/app/shared/models/material.interface.ts
@@ -1,8 +1,7 @@
-import { Entity } from './entity.interface';
-import { EntityType } from './entitytype.enum';
-import { EntityIcon } from './entityicon.enum';
+import { Entity, EntityType, EntityIcon, EntityRoute } from './entity.interface';
 
 export interface Material extends Entity {
     type: EntityType.MATERIAL;
     icon: EntityIcon.MATERIAL;
+    route: EntityRoute.MATERIAL;
 }
