+++ app/src/app/shared/models/motif.interface.ts
@@ -1,8 +1,7 @@
-import { Entity } from './entity.interface';
-import { EntityType } from './entitytype.enum';
-import { EntityIcon } from './entityicon.enum';
+import { Entity, EntityType, EntityIcon, EntityRoute } from './entity.interface';
 
 export interface Motif extends Entity {
     type: EntityType.MOTIF;
     icon: EntityIcon.MOTIF;
+    route: EntityRoute.MOTIF;
 }
