+++ app/src/app/shared/models/movement.interface.ts
@@ -1,9 +1,8 @@
-import { Entity } from './entity.interface';
-import { EntityType } from './entitytype.enum';
-import { EntityIcon } from './entityicon.enum';
+import { Entity, EntityType, EntityIcon, EntityRoute } from './entity.interface';
 
 export interface Movement extends Entity {
     influenced_by: Partial<Entity>[];
     type: EntityType.MOVEMENT;
     icon: EntityIcon.MOVEMENT;
+    route: EntityRoute.MOVEMENT;
 }
